Split models.py into multiple files & directories
