+++ fancypages/models/__init__.py
@@ -1,0 +1,3 @@
+from fancypages.models.base import *
+from fancypages.models.containers import *
+from fancypages.models.widgets import *
