+++ fancypages/models/widgets/__init__.py
@@ -1,0 +1,18 @@
+from fancypages.models.widgets.layouts import (
+    TabWidget,
+    TwoColumnLayoutWidget,
+    ThreeColumnLayoutWidget,
+    FourColumnLayoutWidget,
+)
+from fancypages.models.widgets.content import (
+    TextWidget,
+    TitleTextWidget,
+    ImageWidget,
+)
+from fancypages.models.widgets.product import (
+    SingleProductWidget,
+    HandPickedProductsPromotionWidget,
+    AutomaticProductsPromotionWidget,
+    OfferWidget,
+)
+from fancypages.models.widgets.social import (VideoWidget, TwitterWidget)
